read or write file

+++ updateGoogleDNS.py
@@ -70,13 +70,17 @@ except Exception:
     logging.warn('not found enviroment PATH_INSTALL_SCRIPT_PYTHON')
     
 try:
-    os.path.isfile(PATH_INSTALL+"/my_ip.txt") 
-    logging.info('valid file my_ip.txt')
+    if(os.path.isfile(PATH_INSTALL+"/my_ip.txt") != True):
+        file = open(PATH_INSTALL+"/my_ip.txt", "r")
+        print(file.read())
+    else:
+        file = open(PATH_INSTALL+"/my_ip.txt","w") 
+        file.write(my_ip) 
+        file.close()  
+    
 except Exception:
     logging.warn('not found file my_ip.txt')
-    file = open(PATH_INSTALL+"/my_ip.txt","w") 
-    file.write(my_ip) 
-    file.close()
+    exit()
